changed to floats

+++ lib/tariff.rb
@@ -56,9 +56,8 @@ Explanation:  POWER_USAGE - given annual power consumption in kWh
     if arg1 =~ /^cost$/
       # Cost logic
       if arg_size < 4 && arg2 && arg3
-        # Used integers to reduce complexity of floats - this part needs to be improved only if this app will be called by using floats
-        annual_consumption_power = arg2.to_i
-        annual_consumption_gas = arg3.to_i
+        annual_consumption_power = arg2.to_f
+        annual_consumption_gas = arg3.to_f
 
         annual_cost_hash = {}
 
@@ -109,7 +108,7 @@ Explanation:  POWER_USAGE - given annual power consumption in kWh
         # Used integers to reduce complexity of floats - this part needs to be improved only if this app will be called by using floats
         tariff = arg2.to_s
         fuel_type = arg3.to_s
-        monthly_spend = arg4.to_i
+        monthly_spend = arg4.to_f
 
         unless fuel_type =~ /^(gas|power)$/i
           puts 'Cannot find this fuel type'
